Add (unsigned char) cast to ctype macros ftell->ftello, fseek->fseeko File 'newsize' type int->off_t Add visible (long) cast to fsize() when it called to small one message file and result is assigned to long.

+++ usr.bin/mail/edit.c
@@ -115,7 +115,7 @@ edit1(msgvec, type)
 			size = ftello(otf);
 			mp->m_block = blockof(size);
 			mp->m_offset = boffsetof(size);
-			mp->m_size = fsize(fp);
+			mp->m_size = (long)fsize(fp);
 			mp->m_lines = 0;
 			mp->m_flag |= MODIFY;
 			rewind(fp);
